Moving send_responses_from_output_stream to server_session
